bump undertow

+++ testsuite/standalone/src/test/java/org/jboss/as/test/integration/security/ssl/sni/SNICombinedWithALPNTestCase.java
@@ -18,45 +18,18 @@
 
 package org.jboss.as.test.integration.security.ssl.sni;
 
-import static org.jboss.as.controller.client.helpers.ClientConstants.CONTENT;
-import static org.jboss.as.controller.client.helpers.ClientConstants.DEPLOYMENT;
-import static org.jboss.as.controller.client.helpers.Operations.createAddOperation;
-import static org.jboss.as.controller.client.helpers.Operations.createAddress;
-import static org.jboss.as.controller.client.helpers.Operations.createRemoveOperation;
-
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
-import java.io.FilePermission;
-import java.io.IOException;
-import java.lang.reflect.ReflectPermission;
-import java.math.BigInteger;
-import java.net.URI;
-import java.security.KeyManagementException;
-import java.security.KeyPair;
-import java.security.KeyPairGenerator;
-import java.security.KeyStore;
-import java.security.KeyStoreException;
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
-import java.security.PrivateKey;
-import java.security.SecureRandom;
-import java.security.UnrecoverableKeyException;
-import java.security.cert.CertificateException;
-import java.security.cert.X509Certificate;
-import java.time.Instant;
-import java.time.ZonedDateTime;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Date;
-import java.util.TimeZone;
-import java.util.concurrent.CompletableFuture;
-
-import javax.net.ssl.KeyManagerFactory;
-import javax.net.ssl.SSLContext;
-import javax.net.ssl.TrustManagerFactory;
-import javax.security.auth.x500.X500Principal;
-
+import io.undertow.UndertowOptions;
+import io.undertow.client.ClientCallback;
+import io.undertow.client.ClientConnection;
+import io.undertow.client.ClientExchange;
+import io.undertow.client.ClientRequest;
+import io.undertow.client.UndertowClient;
+import io.undertow.protocols.ssl.UndertowXnioSsl;
+import io.undertow.server.DefaultByteBufferPool;
+import io.undertow.util.Headers;
+import io.undertow.util.Methods;
+import io.undertow.util.Protocols;
+import io.undertow.util.StringReadChannelListener;
 import org.jboss.as.controller.client.OperationBuilder;
 import org.jboss.as.controller.client.helpers.ClientConstants;
 import org.jboss.as.controller.client.helpers.Operations;
@@ -74,12 +47,15 @@ import org.jboss.shrinkwrap.api.exporter.ZipExporter;
 import org.jboss.shrinkwrap.api.spec.JavaArchive;
 import org.junit.Assert;
 import org.junit.Assume;
+import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.wildfly.core.testrunner.ManagementClient;
 import org.wildfly.core.testrunner.ServerSetup;
 import org.wildfly.core.testrunner.ServerSetupTask;
 import org.wildfly.core.testrunner.WildflyTestRunner;
+import org.wildfly.security.x500.GeneralName;
+import org.wildfly.security.x500.cert.SubjectAlternativeNamesExtension;
 import org.wildfly.security.x500.cert.X509CertificateBuilder;
 import org.wildfly.test.undertow.UndertowSSLService;
 import org.wildfly.test.undertow.UndertowSSLServiceActivator;
@@ -89,18 +65,47 @@ import org.xnio.Xnio;
 import org.xnio.XnioWorker;
 import org.xnio.ssl.XnioSsl;
 
-import io.undertow.UndertowOptions;
-import io.undertow.client.ClientCallback;
-import io.undertow.client.ClientConnection;
-import io.undertow.client.ClientExchange;
-import io.undertow.client.ClientRequest;
-import io.undertow.client.UndertowClient;
-import io.undertow.protocols.ssl.UndertowXnioSsl;
-import io.undertow.server.DefaultByteBufferPool;
-import io.undertow.util.Headers;
-import io.undertow.util.Methods;
-import io.undertow.util.Protocols;
-import io.undertow.util.StringReadChannelListener;
+import javax.net.ssl.KeyManagerFactory;
+import javax.net.ssl.SSLContext;
+import javax.net.ssl.TrustManagerFactory;
+import javax.security.auth.x500.X500Principal;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
+import java.io.FilePermission;
+import java.io.IOException;
+import java.lang.reflect.ReflectPermission;
+import java.math.BigInteger;
+import java.net.Inet6Address;
+import java.net.InetAddress;
+import java.net.URI;
+import java.net.UnknownHostException;
+import java.security.KeyManagementException;
+import java.security.KeyPair;
+import java.security.KeyPairGenerator;
+import java.security.KeyStore;
+import java.security.KeyStoreException;
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
+import java.security.PrivateKey;
+import java.security.SecureRandom;
+import java.security.UnrecoverableKeyException;
+import java.security.cert.CertificateException;
+import java.security.cert.X509Certificate;
+import java.time.Instant;
+import java.time.ZonedDateTime;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Date;
+import java.util.List;
+import java.util.TimeZone;
+import java.util.concurrent.CompletableFuture;
+
+import static org.jboss.as.controller.client.helpers.ClientConstants.CONTENT;
+import static org.jboss.as.controller.client.helpers.ClientConstants.DEPLOYMENT;
+import static org.jboss.as.controller.client.helpers.Operations.createAddOperation;
+import static org.jboss.as.controller.client.helpers.Operations.createAddress;
+import static org.jboss.as.controller.client.helpers.Operations.createRemoveOperation;
 
 @RunWith(WildflyTestRunner.class)
 @ServerSetup(SNICombinedWithALPNTestCase.Setup.class)
@@ -121,15 +126,27 @@ public class SNICombinedWithALPNTestCase {
     private static File hostNameKeystore;
     private static File ipKeystore;
 
+    private static boolean assumptionsSatisfied = true;
+
     static class Setup implements ServerSetupTask {
 
         @Override
         public void setup(ManagementClient managementClient) throws Exception {
+            if (!canHostAddressBeTranslated()) {
+                // This is checked by Assumption in the beforeClass() method, but needs to be handled also here because
+                // the setup runs before the beforeClass() method.
+                assumptionsSatisfied = false;
+                return;
+            }
+
+            InetAddress[] addresses = InetAddress.getAllByName(TestSuiteEnvironment.getHttpAddress());
+            String hostname = addresses[0].getHostName();
 
             hostNameKeystore = File.createTempFile("test", ".keystore");
             ipKeystore = File.createTempFile("test", ".keystore");
-            generateFileKeyStore(hostNameKeystore, "localhost");
-            generateFileKeyStore(ipKeystore, "127.0.0.1");
+
+            generateFileKeyStore(hostNameKeystore, hostname, null);
+            generateFileKeyStore(ipKeystore, hostname, addresses);
 
             ModelNode credential = new ModelNode();
             credential.get("clear-text").set("password");
@@ -169,9 +186,9 @@ public class SNICombinedWithALPNTestCase {
             managementClient.executeForResult(modelNode);
 
             modelNode = createAddOperation(createAddress(SNI_SSL_CONTEXT));
-            modelNode.get("default-ssl-context").set("host");
+            modelNode.get("default-ssl-context").set("ip");
             ModelNode hostContextMap = new ModelNode();
-            hostContextMap.get("127.0.0.1").set("ip");
+            hostContextMap.get(hostname).set("host");
             modelNode.get("host-context-map").set(hostContextMap);
             managementClient.executeForResult(modelNode);
 
@@ -192,6 +209,11 @@ public class SNICombinedWithALPNTestCase {
 
         @Override
         public void tearDown(ManagementClient managementClient) throws Exception {
+            if (!assumptionsSatisfied) {
+                // No setup has been done if assumptions failed.
+                return;
+            }
+
             hostNameKeystore.delete();
             ipKeystore.delete();
 
@@ -208,36 +230,47 @@ public class SNICombinedWithALPNTestCase {
         }
     }
 
+    @BeforeClass
+    public static void beforeClass() throws UnknownHostException {
+        Assume.assumeFalse("There is no ALPN implementation in IBM JDK 8 and less; also ALPN-hack that serves" +
+                " as a workaround for other JDKs does not work with IBM JDK.", isIbmJdk() && jdkLessThan9());
+        Assume.assumeTrue("Assuming the test if no resolution for the http address",
+                canHostAddressBeTranslated());
+    }
+
+    private static boolean canHostAddressBeTranslated() throws UnknownHostException {
+        InetAddress address = InetAddress.getByName(TestSuiteEnvironment.getHttpAddress());
+        return !address.getHostName().equals(address.getHostAddress());
+    }
+
     @Test
     public void testSimpleViaHostname() throws Exception {
-        Assume.assumeFalse("There is no ALPN implementation in IBM JDK 8 and less; also ALPN-hack that serves" +
-                        " as a workaround for other JDKs does not work with IBM JDK.", isIbmJdk() && jdkLessThan9());
-
+        InetAddress address = InetAddress.getByName(TestSuiteEnvironment.getHttpAddress());
+        String hostname = address.getHostName();
         XnioSsl ssl = createClientSSL(hostNameKeystore);
         UndertowClient client = UndertowClient.getInstance();
         DefaultByteBufferPool pool = new DefaultByteBufferPool(false, 1024);
-        ClientConnection connection = client.connect(new URI("https", null, "localhost", TestSuiteEnvironment.getHttpPort(), "", null, null), XnioWorker.getContextManager().get(), ssl, pool, OptionMap.create(UndertowOptions.ENABLE_HTTP2, true)).get();
-        performSimpleTest(pool, connection);
+        ClientConnection connection = client.connect(new URI("https", null, hostname, TestSuiteEnvironment.getHttpPort(), "", null, null), XnioWorker.getContextManager().get(), ssl, pool, OptionMap.create(UndertowOptions.ENABLE_HTTP2, true)).get();
+        performSimpleTest(pool, connection, hostname);
     }
 
     @Test
     public void testHttpsViaIp() throws Exception {
-        Assume.assumeFalse("There is no ALPN implementation in IBM JDK 8 and less; also ALPN-hack that serves" +
-                " as a workaround for other JDKs does not work with IBM JDK.", isIbmJdk() && jdkLessThan9());
-
+        InetAddress address = InetAddress.getByName(TestSuiteEnvironment.getHttpAddress());
+        String hostname = address instanceof Inet6Address ? "[" + address.getHostAddress() + "]" : address.getHostAddress();
         XnioSsl ssl = createClientSSL(ipKeystore);
         UndertowClient client = UndertowClient.getInstance();
         DefaultByteBufferPool pool = new DefaultByteBufferPool(false, 1024);
-        ClientConnection connection = client.connect(new URI("https", null, "127.0.0.1", TestSuiteEnvironment.getHttpPort(), "", null, null), XnioWorker.getContextManager().get(), ssl, pool, OptionMap.create(UndertowOptions.ENABLE_HTTP2, true)).get();
-        performSimpleTest(pool, connection);
+        ClientConnection connection = client.connect(new URI("https", null, hostname, TestSuiteEnvironment.getHttpPort(), "", null, null), XnioWorker.getContextManager().get(), ssl, pool, OptionMap.create(UndertowOptions.ENABLE_HTTP2, true)).get();
+        performSimpleTest(pool, connection, hostname);
     }
 
-    private void performSimpleTest(DefaultByteBufferPool pool, ClientConnection connection) throws InterruptedException, java.util.concurrent.ExecutionException {
+    private void performSimpleTest(DefaultByteBufferPool pool, ClientConnection connection, String hostname) throws InterruptedException, java.util.concurrent.ExecutionException {
         ClientRequest cr = new ClientRequest()
                 .setPath("/")
                 .setProtocol(Protocols.HTTP_1_1)
                 .setMethod(Methods.GET);
-        cr.getRequestHeaders().add(Headers.HOST, "localhost");
+        cr.getRequestHeaders().add(Headers.HOST, hostname);
         CompletableFuture<String> future = new CompletableFuture<>();
         connection.sendRequest(cr, new ClientCallback<ClientExchange>() {
             @Override
@@ -289,17 +322,17 @@ public class SNICombinedWithALPNTestCase {
     }
 
 
-    static void generateFileKeyStore(File path, String hostName) {
+    static void generateFileKeyStore(File path, String hostName, InetAddress[] addresses) {
         try {
             KeyPairGenerator keyGen = KeyPairGenerator.getInstance("RSA");
             keyGen.initialize(2048, new SecureRandom());
             KeyPair pair = keyGen.generateKeyPair();
-            X509Certificate cert = generateCertificate(pair, hostName);
+            X509Certificate cert = generateCertificate(pair, hostName, addresses);
 
             KeyStore keyStore = KeyStore.getInstance(KeyStore.getDefaultType());
             keyStore.load(null, PASSWORD.toCharArray());
 
-            //Generate self signed certificate
+            //Generate self-signed certificate
             X509Certificate[] chain = new X509Certificate[1];
             chain[0] = cert;
             keyStore.setKeyEntry(ALIAS, pair.getPrivate(), PASSWORD.toCharArray(), chain);
@@ -318,7 +351,7 @@ public class SNICombinedWithALPNTestCase {
         }
     }
 
-    static X509Certificate generateCertificate(KeyPair pair, String hostName) throws Exception {
+    static X509Certificate generateCertificate(KeyPair pair, String hostName, InetAddress[] addresses) throws Exception {
         PrivateKey privkey = pair.getPrivate();
         X509CertificateBuilder builder = new X509CertificateBuilder();
         Date from = new Date();
@@ -335,6 +368,14 @@ public class SNICombinedWithALPNTestCase {
         builder.setVersion(3);
         builder.setSignatureAlgorithmName(SHA_256_WITH_RSA);
         builder.setSigningKey(privkey);
+        List<GeneralName> subjectAlternativeNames = new ArrayList<>();
+        subjectAlternativeNames.add(new GeneralName.DNSName(hostName));
+        if (addresses != null) {
+            for (int i = 0; i < addresses.length; i++) {
+                subjectAlternativeNames.add(new GeneralName.IPAddress(addresses[i].getHostAddress()));
+            }
+        }
+        builder.addExtension(new SubjectAlternativeNamesExtension(false, subjectAlternativeNames));
         return builder.build();
     }
 
@@ -367,11 +408,7 @@ public class SNICombinedWithALPNTestCase {
      * @return IbmX509 in case of IBM JDK, SunX509 otherwise
      */
     private static String keyAlgorithm() {
-        if (isIbmJdk()) {
-            return "IbmX509";
-        } else {
-            return "SunX509";
-        }
+        return "SunX509";
     }
 
     /**
